requirements.txt
# LogiQ AI - Requirements
# Python 3.10+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Processing
pandas==2.1.3
sqlalchemy==2.0.23

# Frontend
streamlit==1.28.2

# HTTP Client
requests==2.31.0

# Testing
pytest==7.4.3

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# AI/ML
google-generativeai==0.3.1
openai==1.3.7

# Optional: BigQuery support
# google-cloud-bigquery==3.13.0
